调整最终选举时的数据结构

+++ consensus/snapshots/cad_node_snap.go
@@ -22,6 +22,7 @@ import (
 	//"sort"
 	//"fmt"
 	"encoding/json"
+	//"math/big"
 	
 	"github.com/hpb-project/ghpb/common"
 	//"github.com/hpb-project/ghpb/core/types"
@@ -48,7 +49,7 @@ type CadNodeSnap struct {
 type CadWinner struct {
 	NetworkId     string `json:"networkid"`             // 获胜者的网络ID
 	Address       string `json:"address"`      // 获胜者的地址
-	VoteIndex     float64 `json:"voteIndex"`       // 获胜者的地址
+	VoteIndex     uint64 `json:"voteIndex"`       // 获胜者的地址
 }
 
 // 创建对象
